Avoid exposing internal types by improving internal organization.
